Factor out Retry Rule
